Skip sending packets as long as GPS isn't fix yet
This moves the DccPassthrough into its own file.

Additionally, it enables mib.itsGnLocalGnAddr.is_manually_configured(true); to make socktap work on the loopback interface.
